% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kames2016.r
\docType{data}
\name{kames2016}
\alias{kames2016}
\title{Kames Bay survey data 2016}
\format{
An object of class \code{data.frame} with 16 rows and 7 columns.
}
\source{
Students and staff of Population and Community Ecology module 2016
}
\usage{
data(kames2016)
}
\description{
Data from box core survey of Kames Bay, Millport, Great Cumbrae, Scotland.
Four replicate box core samples were taken at 5m intervals and sieved through
a 0.5mm sieve before identification.  Only partial data (first 4 stations)
are included.
}
\keyword{datasets}
