% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kames2017.r
\docType{data}
\name{kames2017}
\alias{kames2017}
\title{Kames Bay survey data 2017}
\format{
An object of class \code{data.frame} with 67 rows and 10 columns.
}
\source{
Students and staff of ENV09103 Population and Community Ecology
module 2017
}
\usage{
data(kames2017)
}
\description{
Data from box core survey of Kames Bay, Millport, Great Cumbrae, Scotland.
Four replicate box core samples were taken at 5m or 10m intervals and sieved
through a 0.5mm sieve before identification.
}
\keyword{datasets}
