% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kites.r
\name{plot_kite}
\alias{plot_kite}
\title{Plot kite diagram}
\usage{
plot_kite(survey, min_abund = 50)
}
\arguments{
\item{survey}{Survey data.}

\item{min_abund}{Minimum total abundance of taxa to be included on plot.}
}
\value{
Kite diagram plot (base graphics).
}
\description{
Plot kite diagram
}
